+++ debs/build.sh
@@ -21,7 +21,8 @@ debug=$7
 checksum=$8
 wazuh_packages_branch=$9
 use_local_specs=${10}
-local_source_code=${11}
+future=${11}
+local_source_code=${12}
 
 if [ -z "${package_release}" ]; then
     package_release="1"
@@ -50,10 +51,29 @@ cp -R wazuh* ${build_dir}/${build_target}/wazuh-${build_target}-${wazuh_version}
 if [ "${use_local_specs}" = "no" ]; then
     curl -sL https://github.com/wazuh/wazuh-packages/tarball/${wazuh_packages_branch} | tar zx
     package_files="wazuh*/debs"
-    specs_path="${package_files}/SPECS"
+    specs_path=$(find . -type d -name "SPECS" -path "*debs*")
 else
     package_files="/specs"
     specs_path="${package_files}/SPECS"
+fi
+
+if [[ "${future}" == "yes" ]]; then
+    # MODIFY VARIABLES
+    base_version=$wazuh_version
+    MAJOR=$(echo $base_version | cut -dv -f2 | cut -d. -f1)
+    MINOR=$(echo $base_version | cut -d. -f2)
+    wazuh_version="${MAJOR}.30.0"
+    file_name="wazuh-${build_target}-${wazuh_version}-${package_release}"
+    old_name="wazuh-${build_target}-${base_version}-${package_release}"
+    package_full_name=wazuh-${build_target}-${wazuh_version}
+    old_package_name=wazuh-${build_target}-${base_version}
+    sources_dir="${build_dir}/${build_target}/${package_full_name}"
+
+    # PREPARE FUTURE SPECS AND SOURCES
+    cp -r "${specs_path}/${base_version}" "${specs_path}/${wazuh_version}"
+    mv ${build_dir}/${build_target}/${old_package_name} ${build_dir}/${build_target}/${package_full_name}
+    find "${build_dir}/${package_name}" "${specs_path}/${wazuh_version}" \( -name "*VERSION*" -o -name "*changelog*" \) -exec sed -i "s/${base_version}/${wazuh_version}/g" {} \;
+    sed -i "s/\$(VERSION)/${MAJOR}.${MINOR}/g" "${build_dir}/${build_target}/${package_full_name}/src/Makefile"
 fi
 cp -pr ${specs_path}/${wazuh_version}/wazuh-${build_target}/debian ${sources_dir}/debian
 cp -p ${package_files}/gen_permissions.sh ${sources_dir}
@@ -85,11 +105,11 @@ mk-build-deps -ir -t "apt-get -o Debug::pkgProblemResolver=yes -y"
 # Build package
 if [[ "${architecture_target}" == "amd64" ]] ||  [[ "${architecture_target}" == "ppc64le" ]] || \
     [[ "${architecture_target}" == "arm64" ]]; then
-    debuild -eLD_LIBRARY_PATH --prepend-path="/usr/local/gcc-5.5.0/bin" -b -uc -us
+    debuild --rootcmd=sudo -eLD_LIBRARY_PATH --prepend-path="/usr/local/gcc-5.5.0/bin" -b -uc -us
 elif [[ "${architecture_target}" == "armhf" ]]; then
-    linux32 debuild -eLD_LIBRARY_PATH --prepend-path="/usr/local/gcc-5.5.0/bin" -b -uc -us
+    linux32 debuild --rootcmd=sudo -eLD_LIBRARY_PATH --prepend-path="/usr/local/gcc-5.5.0/bin" -b -uc -us
 else
-    linux32 debuild -eLD_LIBRARY_PATH --prepend-path="/usr/local/gcc-5.5.0/bin" -ai386 -b -uc -us
+    linux32 debuild --rootcmd=sudo -eLD_LIBRARY_PATH --prepend-path="/usr/local/gcc-5.5.0/bin" -ai386 -b -uc -us
 fi
 
 deb_file="wazuh-${build_target}_${wazuh_version}-${package_release}_${architecture_target}.deb"
@@ -98,5 +118,4 @@ pkg_path="${build_dir}/${build_target}"
 if [[ "${checksum}" == "yes" ]]; then
     cd ${pkg_path} && sha512sum ${deb_file} > /var/local/checksum/${deb_file}.sha512
 fi
-
 mv ${pkg_path}/${deb_file} /var/local/wazuh