+++ ova/Ova2Ovf.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 # Copyright (C) 2019, Wazuh Inc.
 #
-# Ova2Ovf.py        Helper script to convert VBox .ova export 
+# Ova2Ovf.py        Helper script to convert VBox .ova export
 #                   for import to VMWare ESXi
 #
 # Original author: eshizhan https://github.com/eshizhan
-# Author: Neova Health 
+# Author: Neova Health
 # forked from : https://gist.github.com/eshizhan/6650285
 # Modified by Wazuh, Inc
 
@@ -62,4 +62,3 @@ with open(ovaF, 'wb') as nfp:
 with tarfile.open(destfile, "w") as t:
     for name in ovaFiles:
         t.add(name)
-