+++ ova/provision.sh
@@ -4,17 +4,36 @@ set -exf
 repo_branch=$(echo "$1" | cut -c1-3)
 repo_baseurl=$(echo "$1" | cut -c1-2)
 WAZUH_VERSION=$1
-ELK_VERSION=$2
-STATUS_PACKAGES=$3
-DIRECTORY=$4
+OPENDISTRO_VERSION=$2
+ELK_VERSION=$3
+PACKAGES_REPOSITORY=$4
+BRANCH=$5
+UI_REVISION=$6
+DIRECTORY="/var/ossec"
 ELK_MAJOR=`echo ${ELK_VERSION}|cut -d"." -f1`
 ELK_MINOR=`echo ${ELK_VERSION}|cut -d"." -f2`
-config_files="/vagrant/Config_files"
 
-. /vagrant/Libraries/wazuh_functions.sh
-. /vagrant/Libraries/elastic_functions.sh
+CURRENT_PATH="$( cd $(dirname $0) ; pwd -P )"
+config_files="${CURRENT_PATH}/Config_files"
+automatic_set_ram_location="/etc/"
+libraries_files="${CURRENT_PATH}/Libraries/"
+
+echo "${PACKAGES_REPOSITORY}"
+
+
+. ${CURRENT_PATH}/Libraries/provision-opendistro.sh
+
 
 # Setting wazuh default root password
+
+cp ${libraries_files}/automatic_set_ram.sh ${automatic_set_ram_location}
+chmod +x "${automatic_set_ram_location}/automatic_set_ram.sh"
+echo "@reboot . /etc/automatic_set_ram.sh" >> ram_cron
+
+
+crontab ram_cron
+rm -rf ram_cron
+
 yes wazuh | passwd root
 hostname wazuhmanager
 
@@ -25,13 +44,18 @@ echo "PermitRootLogin yes" >> /etc/ssh/sshd_config
 # Dependences
 yum install openssl -y
 
-install_wazuh
-
-elastic_stack_${ELK_MAJOR}
+installPrerequisites
+addWazuhrepo
+addElasticRepo
+installWazuh
+installElasticsearch
+installFilebeat
+installKibana
+configWazuh
+checkInstallation
+cleanInstall
 
 rm -rf /vagrant
 
-systemctl stop kibana
-systemctl stop  elasticsearch
-systemctl stop wazuh-manager
-systemctl stop wazuh-api
+systemctl stop kibana filebeat elasticsearch
+systemctl enable wazuh-manager