+++ rpms/CentOS/7/armv7hl/build_deps.sh
@@ -313,4 +313,6 @@ ${rpm_url}/gl-manpages-1.1-7.20130122.el7.noarch.rpm \
 ${rpm_url}/libICE-1.0.9-9.el7.armv7hl.rpm \
 ${rpm_url}/libICE-devel-1.0.9-9.el7.armv7hl.rpm \
 ${rpm_url}/wget-1.14-18.el7_6.1.armv7hl.rpm \
-${rpm_url}/libcurl-devel-7.29.0-57.el7.armv7hl.rpm
+${rpm_url}/libcurl-devel-7.29.0-57.el7.armv7hl.rpm \
+${rpm_url}/cmake3-data-3.17.3-3.el7.noarch.rpm \
+${rpm_url}/cmake3-3.17.3-3.el7.armv7hl.rpm