+++ CHANGELOG.md
@@ -24,5 +24,5 @@ All notable changes to this project will be documented in this file.
 
 ### Removed
 
-- Removed Wazuh API package has been removed. Now, the Wazuh API is embedded into the Wazuh Manager installation ([wazuh/wazuh#5721](https://github.com/wazuh/wazuh/pull/5721)).
+- Wazuh API package has been removed. Now, the Wazuh API is embedded into the Wazuh Manager installation ([wazuh/wazuh#5721](https://github.com/wazuh/wazuh/pull/5721)).
 - Removed OpenSCAP files and integration ([#466](https://github.com/wazuh/wazuh-packages/pull/466)).